[level 2] Title: 피보나치 수, Time: 2.48 ms, Memory: 72.2 MB -BaekjoonHub

+++ 프로그래머스/2/12945. 피보나치 수/피보나치 수.java
@@ -2,7 +2,7 @@ class Solution {
     public int solution(int n) {
         int answer = 0;
         
-        long[] arr = new long[n + 1];
+        int[] arr = new int[n + 1];
         arr[0] = 0;
         arr[1] = 1;
